Use MessageView in search tab
